+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 - Added OIDC activation [(#7)](https://github.com/wultra/mtoken-sdk-flutter/issues/7)
 - Added support for UserOperation attribute type Alert [(#15)](https://github.com/wultra/mtoken-sdk-flutter/issues/15)
 - Added Pre-Approval Screens and WMTPreApprovalScreensRecorder/Mobile Token data support [(#27)](https://github.com/wultra/mtoken-sdk-flutter/issues/27)
+- Improved proximity check time handling, timestamps are now adjusted to server-synchronized time automatically during authorization [(#36)](https://github.com/wultra/mtoken-sdk-flutter/issues/36)
 
 ## 1.0.0